Pull jemalloc compat things into a separate file.
