+++ src/test/func/malloc/malloc.cc
@@ -339,14 +339,6 @@ int main(int argc, char** argv)
     abort();
   }
 
-#ifndef __PIC__
   snmalloc::debug_check_empty<snmalloc::Globals>();
-  void* bootstrap = __je_bootstrap_malloc(42);
-  if (bootstrap == nullptr)
-  {
-    printf("Failed to allocate from bootstrap malloc\n");
-  }
-  __je_bootstrap_free(bootstrap);
-#endif
   return 0;
 }
